refactor: separate UI/REPL from commands and extract constants
